refactor : 네이밍 직관적으로 수정
